Migrate to null safety.

+++ lib/widgets/synced_photos_screen/synced_photos_screen.dart
@@ -6,20 +6,20 @@ import 'package:photos_sync/widgets/common/confirm_dialog.dart';
 import 'package:provider/provider.dart';
 
 class SyncedPhotosScreen extends StatefulWidget {
-  const SyncedPhotosScreen({Key key}) : super(key: key);
+  const SyncedPhotosScreen({Key? key}) : super(key: key);
 
   @override
   _SyncedPhotosScreenState createState() => _SyncedPhotosScreenState();
 }
 
 class _SyncedPhotosScreenState extends State<SyncedPhotosScreen> {
-  Future<List<SyncedPhoto>> _syncedPhotosFuture;
+  Future<List<SyncedPhoto>>? _syncedPhotosFuture;
 
   @override
   void initState() {
     super.initState();
 
-    _syncedPhotosFuture = context.read<IDatabaseService>().getPhotos(user: UserPreferences.getUsername());
+    _syncedPhotosFuture = context.read<IDatabaseService>().getPhotos(user: UserPreferences.getUsername()!);
   }
 
   @override
@@ -30,19 +30,20 @@ class _SyncedPhotosScreenState extends State<SyncedPhotosScreen> {
           PopupMenuButton<int>(
             icon: Icon(Icons.more_vert),
             onSelected: (_) async {
-              final shouldDeleteAll = await showConfirmDialog(
+              final shouldDeleteAll = await (showConfirmDialog(
                 context: context,
                 title: I18n.syncAllPhotosPopupTitleText,
                 description: I18n.syncAllPhotosPopupDescriptionText,
-              );
-              if (shouldDeleteAll) {
-                final photos = await context.read<IDatabaseService>().getPhotos(user: UserPreferences.getUsername());
+              ));
+              if (shouldDeleteAll != null && shouldDeleteAll) {
+                final photos = await context.read<IDatabaseService>().getPhotos(user: UserPreferences.getUsername()!);
                 for (final photo in photos) {
                   await context.read<ISyncService>().deleteFile(photo);
                 }
 
                 setState(() {
-                  _syncedPhotosFuture = context.read<IDatabaseService>().getPhotos(user: UserPreferences.getUsername());
+                  _syncedPhotosFuture =
+                      context.read<IDatabaseService>().getPhotos(user: UserPreferences.getUsername()!);
                 });
               }
             },
@@ -66,15 +67,15 @@ class _SyncedPhotosScreenState extends State<SyncedPhotosScreen> {
               );
             default:
               if (snapshot.connectionState == ConnectionState.done && snapshot.hasData) {
-                snapshot.data.sort((a, b) => a.absoluteFilepath.compareTo(b.absoluteFilepath));
+                snapshot.data!.sort((a, b) => a.absoluteFilepath!.compareTo(b.absoluteFilepath!));
                 return ListView.builder(
-                  itemCount: snapshot.data.length,
+                  itemCount: snapshot.data!.length,
                   itemBuilder: (_, index) => Dismissible(
-                    key: Key(snapshot.data[index].absoluteFilepath),
+                    key: Key(snapshot.data![index].absoluteFilepath!),
                     direction: DismissDirection.endToStart,
                     child: ListTile(
-                      title: Text(snapshot.data[index].filename),
-                      subtitle: Text(snapshot.data[index].absoluteFilepath),
+                      title: Text(snapshot.data![index].filename!),
+                      subtitle: Text(snapshot.data![index].absoluteFilepath!),
                     ),
                     background: Container(
                       alignment: AlignmentDirectional.centerEnd,
@@ -87,7 +88,7 @@ class _SyncedPhotosScreenState extends State<SyncedPhotosScreen> {
                         ),
                       ),
                     ),
-                    onDismissed: (_) => context.read<ISyncService>().deleteFile(snapshot.data[index]),
+                    onDismissed: (_) => context.read<ISyncService>().deleteFile(snapshot.data![index]),
                     confirmDismiss: (_) async => await showConfirmDialog(
                       context: context,
                       title: I18n.syncSinglePhotoPopupTitleText,
